txscript: Move stake votebits script to consensus.
This moves the GenerateSSGenVotes method to script.go since it is used
in consensus code and therefore should not be in standard.go which is
only intended for standardness code.

It also moves the associated test to script_test.go to match and does
some other misc cleanup such as improving the comments and tests for
consistency.
